+++ src/sdk/cpp/examples/CMakeLists.txt
@@ -1,7 +1,5 @@
-add_subdirectory(event_consumer)
-add_subdirectory(function_provider)
-add_subdirectory(echo)
-add_subdirectory(callback_mode)
-add_subdirectory(subclass_mode)
 add_subdirectory(basic)
+add_subdirectory(callback_mode)
+add_subdirectory(echo)
 add_subdirectory(functions)
+add_subdirectory(subclass_mode)
