+++ src/sdk/cpp/examples/basic/basic.cpp
@@ -12,7 +12,7 @@
 #include <iostream>
 #include "nlohmann/json.hpp"
 
-#include "iotea.hpp"
+#include "client.hpp"
 #include "mqtt_client.hpp"
 
 using json = nlohmann::json;
