+++ src/sdk/cpp/examples/callback_mode/callback_mode.cpp
@@ -12,7 +12,7 @@
 #include <iostream>
 #include "nlohmann/json.hpp"
 
-#include "iotea.hpp"
+#include "client.hpp"
 
 using json = nlohmann::json;
 using iotea::core::Service;
