+++ src/sdk/cpp/examples/echo/echo_consumer.cpp
@@ -13,10 +13,9 @@
 #include <memory>
 
 #include "nlohmann/json.hpp"
-#include "iotea.hpp"
+#include "client.hpp"
 #include "logging.hpp"
 #include "mqtt_client.hpp"
-#include "schema.hpp"
 
 using namespace iotea::core;
 using json = nlohmann::json;
