+++ src/sdk/cpp/examples/echo/echo_observer.cpp
@@ -13,7 +13,7 @@
 #include <memory>
 
 #include "nlohmann/json.hpp"
-#include "iotea.hpp"
+#include "client.hpp"
 #include "logging.hpp"
 #include "mqtt_client.hpp"
 #include "schema.hpp"
