+++ src/sdk/cpp/examples/echo/echo_provider.cpp
@@ -13,7 +13,7 @@
 #include <string>
 
 #include "nlohmann/json.hpp"
-#include "iotea.hpp"
+#include "client.hpp"
 #include "logging.hpp"
 #include "mqtt_client.hpp"
