+++ src/sdk/cpp/examples/subclass_mode/subclass_mode.cpp
@@ -9,10 +9,9 @@
  ****************************************************************************/
 
 #include <csignal>
-#include <iostream>
 #include "nlohmann/json.hpp"
 
-#include "iotea.hpp"
+#include "client.hpp"
 #include "mqtt_client.hpp"
 
 using json = nlohmann::json;
