+++ src/sdk/cpp/lib/include/schema.hpp
@@ -8,8 +8,8 @@
  * SPDX-License-Identifier: MPL-2.0
  ****************************************************************************/
 
-#ifndef IOTEA_SCHEMA_HPP
-#define IOTEA_SCHEMA_HPP
+#ifndef SRC_SDK_CPP_LIB_INCLUDE_SCHEMA_HPP_
+#define SRC_SDK_CPP_LIB_INCLUDE_SCHEMA_HPP_
 
 #include <memory>
 #include <regex>
@@ -772,4 +772,4 @@ schema::rule_ptr NelsonOut3Se(const std::string& feature, const std::string& typ
 }  // namespace core
 }  // namespace iotea
 
-#endif // IOTEA_SCHEMA_HPP_
+#endif // SRC_SDK_CPP_LIB_INCLUDE_SCHEMA_HPP_
