+++ src/sdk/cpp/lib/src/util.cpp
@@ -11,7 +11,6 @@
 #include <chrono>
 #include <random>
 #include <iomanip>
-#include <string>
 #include <sstream>
 
 #include "util.hpp"
