+++ test/integration-tests/testset_sdk/cpp/function_provider.cpp
@@ -12,7 +12,7 @@
 #include <memory>
 
 #include "nlohmann/json.hpp"
-#include "iotea.hpp"
+#include "client.hpp"
 
 using json = nlohmann::json;
 using namespace iotea::core;
