+++ test/integration-tests/testset_sdk/cpp/testset_sdk.cpp
@@ -12,8 +12,8 @@
 #include <memory>
 
 #include "nlohmann/json.hpp"
-#include "iotea.hpp"
 #include "talent_test.hpp"
+#include "client.hpp"
 
 using json = nlohmann::json;
 using namespace iotea::core;
